+++ src/main/java/com/lucendar/strm/common/StrmMsgs.java
@@ -12,11 +12,11 @@ public class StrmMsgs {
     public static final int STRM_MSG__LiveStrmCtrlReq = 1;
     public static final int STRM_MSG__ReplayStrmCtrlReq = 2;
     public static final int STRM_MSG__CheckCtrlReq = 3;
-    public static final int STRM_MSG__OpenChannelReq = 4;
-    public static final int STRM_MSG__KeepChannelReq = 5;
-    public static final int STRM_MSG__ReleaseChannelsReq = 6;
+    public static final int STRM_MSG__OpenStrmReq = 4;
+    public static final int STRM_MSG__KeepStrmReq = 5;
+    public static final int STRM_MSG__ReleaseStrmsReq = 6;
     public static final int STRM_MSG__QryServerStatusReq = 7;
-    public static final int STRM_MSG__CheckChannelReq = 8;
+    public static final int STRM_MSG__CheckStrmReq = 8;
     public static final int STRM_MSG__ClosedStrm = 9;
     public static final int STRM_MSG__QryRecentClosedStrmReq = 10;
     public static final int STRM_MSG__StoredAv = 11;