+++ src/main/java/com/lucendar/strm/common/strm/KeepStrmReq.java
@@ -6,14 +6,14 @@ import com.lucendar.strm.common.StrmMsgs;
 import java.util.Arrays;
 import java.util.StringJoiner;
 
-public class KeepChannelReq implements StrmMsg {
+public class KeepStrmReq implements StrmMsg {
     private String[] reqIds;
     private ServerHint strmServerHint;
 
-    public KeepChannelReq() {
+    public KeepStrmReq() {
     }
 
-    public KeepChannelReq(String[] reqIds, ServerHint strmServerHint) {
+    public KeepStrmReq(String[] reqIds, ServerHint strmServerHint) {
         this.reqIds = reqIds;
         this.strmServerHint = strmServerHint;
     }
@@ -37,12 +37,12 @@ public class KeepChannelReq implements StrmMsg {
     @Override
     public String toString() {
         if (strmServerHint != null)
-            return new StringJoiner(", ", KeepChannelReq.class.getSimpleName() + "[", "]")
+            return new StringJoiner(", ", KeepStrmReq.class.getSimpleName() + "[", "]")
                     .add("reqIds=" + Arrays.toString(reqIds))
                     .add("strmServerHint=" + strmServerHint)
                     .toString();
         else
-            return new StringJoiner(", ", KeepChannelReq.class.getSimpleName() + "[", "]")
+            return new StringJoiner(", ", KeepStrmReq.class.getSimpleName() + "[", "]")
                     .add("reqIds=" + Arrays.toString(reqIds))
                     .toString();
 
@@ -50,6 +50,6 @@ public class KeepChannelReq implements StrmMsg {
 
     @Override
     public int msgId() {
-        return StrmMsgs.STRM_MSG__KeepChannelReq;
+        return StrmMsgs.STRM_MSG__KeepStrmReq;
     }
 }